+++ internal/runtime/executor/claude_fingerprint_policy.go
@@ -1,11 +1,13 @@
 package executor
 
 import (
+	"fmt"
 	"strings"
 	"sync"
 
 	claudeauth "github.com/router-for-me/CLIProxyAPI/v7/internal/auth/claude"
 	"github.com/router-for-me/CLIProxyAPI/v7/internal/config"
+	"github.com/router-for-me/CLIProxyAPI/v7/internal/runtime/executor/helps"
 	cliproxyauth "github.com/router-for-me/CLIProxyAPI/v7/sdk/cliproxy/auth"
 	log "github.com/sirupsen/logrus"
 )
@@ -108,4 +110,32 @@ func resolveClaudeFingerprintPolicy(cfg *config.Config, auth *cliproxyauth.Auth,
 		InjectDiagnostics:    profileClaudeCodeCLI,
 		OAuthCancellation:    authIsOAuth,
 	}
+}
+
+// applyClaudeCLIIdentity applies the Claude Code CLI credential identity to the
+// upstream Messages body. It is the single implementation behind both the
+// streaming and the non-streaming request paths; keep it that way.
+//
+// ApplyCLIIdentity and ProfileClaudeCodeCLI are the same predicate, so
+// sessionID has already been resolved by ClaudeAgentSessionUUIDForRequest,
+// which always returns a UUID. Do not add a second session source here: a
+// per-apiKey cached ID would silently break agent-conversation continuity.
+//
+// API keys seed the synthesized identity from the key itself; delegated
+// providers such as Kimi seed from the stable auth identity, so an access-token
+// rotation does not rotate the device fingerprint.
+func applyClaudeCLIIdentity(body []byte, auth *cliproxyauth.Auth, apiKey, upstreamURL, sessionID string, synthesize bool) ([]byte, error) {
+	identitySeed := apiKey
+	if isKimiMessagesUpstream(auth, upstreamURL) {
+		identitySeed = helps.ClaudeCLIAuthIdentitySeed(auth)
+	}
+	identityAuth, errIdentity := helps.PrepareClaudeCLIFingerprintAuth(auth, identitySeed, synthesize)
+	if errIdentity != nil {
+		return nil, fmt.Errorf("ensure Claude CLI fingerprint identity: %w", errIdentity)
+	}
+	updated, _, errApply := helps.ApplyClaudeCredentialMetadata(body, identityAuth, sessionID)
+	if errApply != nil {
+		return nil, fmt.Errorf("apply Claude credential metadata: %w", errApply)
+	}
+	return updated, nil
 }
